feature add address bus trait for easing the development of the cpu (#59)
* add addressbus trait

* correct error

* extract address bus

* correct doc test

* extract address

* expose bus trait

* format lib.rs

* correct failing compilation

* extract component in their own module

* rework Bus trait

* update trait

* correct error

* correct mbc5 crate error

* implement u16
