+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
   "gb-roms",
   "gb-ppu",
   "gb-lcd",
+  "gb-cpu",
   "gb-dbg",
-  "gb-cpu"
+  "gb-bus"
 ]
