+++ gb-bus/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "gb-cpu"
+name = "gb-bus"
 version = "0.1.0"
-authors = ["firelight.flagboy <firelight.flagboy@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+rand = { version = "0.8.4", features = ["small_rng"] }
