+++ gb-bus/src/address_bus/error.rs
@@ -1,0 +1,7 @@
+use crate::Address;
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum Error {
+    BusError(Address),
+    SegmentationFault(Address),
+}
