+++ gb-cpu/src/lib.rs
@@ -1,5 +1,1 @@
-pub mod address_bus;
-
 pub mod bus;
-
-pub use address_bus::{AddressBus, Area, FileOperation};
