+++ gb-roms/Cargo.toml
@@ -11,4 +11,4 @@ clap = { version = "3.0.0-beta.2" }
 
 [dependencies]
 modular-bitfield = "0.11.2"
-gb-cpu = { path = "../gb-cpu" }
+gb-bus = { path = "../gb-bus" }
